requirements.txt
openai>=1.0.0
websocket-client>=1.6.0
aiohttp>=3.8.0
asyncio
requests>=2.28.0
pyyaml>=6.0